rating

+++ src/main/java/com/PI/apiBooking/Repository/IRatingRepository.java
@@ -11,5 +11,5 @@ import java.util.Optional;
 public interface IRatingRepository extends JpaRepository<Rating, Long> {
 
     @Query("SELECT AVG(r.score) FROM Rating r WHERE product.id = ?1")
-    Optional<Double> buscarPorProducto(Long productId);
+    Optional<Integer> buscarPorProducto(Long productId);
 }
